requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
gql>=3.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
numpy>=1.24.0
tqdm>=4.65.0
slowapi>=0.1.8
python-multipart>=0.0.6
aiohttp>=3.8.0
typing-extensions>=4.0.0
setuptools>=68.0.0
wheel>=0.40.0